# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Donation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /eye-calls:
    get:
      operationId: listEyeCalls
      tags: [eye-calls]
      summary: List eye donation calls (filtered by role/unit)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, team_sent, completed, cancelled]
        - name: unitId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of eye calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCallList"
    post:
      operationId: createEyeCall
      tags: [eye-calls]
      summary: Create a new eye donation call (requires auth for internal, public uses separate endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EyeCallInput"
      responses:
        "201":
          description: Created eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "400":
          description: Invalid input
  /eye-calls/{id}:
    get:
      operationId: getEyeCall
      tags: [eye-calls]
      summary: Get a specific eye call by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eye call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found
    patch:
      operationId: updateEyeCall
      tags: [eye-calls]
      summary: Update an eye call (status, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EyeCallUpdate"
      responses:
        "200":
          description: Updated eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found
    delete:
      operationId: deleteEyeCall
      tags: [eye-calls]
      summary: Delete an eye call (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Eye call not found
  /eye-calls/{id}/status:
    patch:
      operationId: updateEyeCallStatus
      tags: [eye-calls]
      summary: Update eye call status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found

  # ─── PUBLIC FORM ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    EyeCall:
      type: object
      required: [id, callId, donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId, unitName, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        callId:
          type: string
        donorName:
          type: string
        donorAge:
          type: integer
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
        referrerMobile:
          type: string
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
        unitName:
          type: string
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: ["string", "null"]
        thirdPartyHospitalDetails:
          type: ["string", "null"]
        thirdPartyHelperContact:
          type: ["string", "null"]
        attendedByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EyeCallList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EyeCall"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EyeCallInput:
      type: object
      required: [donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId]
      properties:
        donorName:
          type: string
          minLength: 1
        donorAge:
          type: integer
          minimum: 0
          maximum: 120
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
          minLength: 1
        referrerMobile:
          type: string
          minLength: 10
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
        notes:
          type: string
        thirdPartyHospitalDetails:
          type: string
        thirdPartyHelperContact:
          type: string
        attendedByUserId:
          type: integer
    EyeCallUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: string
        thirdPartyHospitalDetails:
          type: string
        thirdPartyHelperContact:
          type: string
        attendedByUserId:
          type: integer
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: string
...